Take out HintHandler out of StatePlayerRecyclerViewAssembler
