+++ app/src/main/java/org/oppia/android/app/player/state/hintsandsolution/HintHandlerProdModule.kt
@@ -1,0 +1,10 @@
+package org.oppia.android.app.player.state.hintsandsolution
+
+import dagger.Binds
+import dagger.Module
+
+@Module
+interface HintHandlerProdModule {
+  @Binds
+  fun bindsHintHandler(prodHintHandler: ProdHintHandler): HintHandler
+}
